Made the window options class in export a superclass for multiple windows
